@@ -3,16 +3,13 @@
   "version": "0.0.3",
   "description": "======= Bookery =======",
   "main": "dist/electron.js",
-  "directories": {
-    "doc": "docs",
-    "test": "tests"
-  },
   "scripts": {
-    "build": "webpack",
     "test": "mocha",
-    "start": "electron .",
-    "prestart": "npm run build",
-    "lint": "eslint --ext=js,jsx,ts,tsx ."
+    "lint": "eslint --ext=js,jsx,ts,tsx .",
+    "dev": "electron-webpack dev",
+    "compile": "electron-webpack",
+    "dist": "npm run compile && electron-builder",
+    "postinstall": "electron-builder install-app-deps"
   },
   "repository": {
     "type": "git",
@@ -25,47 +22,45 @@
   },
   "homepage": "https://github.com/0916dhkim/bookery#readme",
   "dependencies": {
-    "@types/jquery": "^3.3.32",
-    "@types/react": "^16.9.20",
+    "@types/jquery": "^3.3.33",
+    "@types/react": "^16.9.23",
     "@types/react-dom": "^16.9.5",
-    "@types/webpack": "^4.41.6",
-    "@types/webpack-merge": "^4.1.5",
-    "css-loader": "^3.4.2",
-    "electron": "^8.0.1",
-    "file-loader": "^5.1.0",
-    "fuse.js": "^3.4.6",
+    "fuse.js": "^3.6.1",
     "immer": "^5.3.6",
     "jquery": "^3.4.1",
     "moment": "^2.24.0",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "react": "^16.13.0",
+    "react-dom": "^16.13.0",
     "semantic-ui-css": "^2.4.1",
     "semantic-ui-react": "^0.88.2",
-    "style-loader": "^1.1.3",
-    "ts-loader": "^6.2.1",
-    "typescript": "^3.7.5",
-    "url-loader": "^3.0.0",
-    "webpack": "^4.41.6",
-    "webpack-cli": "^3.3.11",
-    "webpack-merge": "^4.2.2"
+    "source-map-support": "^0.5.16",
+    "typescript": "^3.8.3",
+    "webpack": "^4.42.0"
   },
   "devDependencies": {
-    "@testing-library/react": "^9.4.0",
+    "@babel/preset-react": "^7.8.3",
+    "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^10.0.0",
-    "@types/mocha": "^7.0.1",
+    "@types/mocha": "^7.0.2",
     "@types/sinon": "^7.5.2",
-    "@typescript-eslint/eslint-plugin": "^2.20.0",
-    "@typescript-eslint/parser": "^2.20.0",
+    "@typescript-eslint/eslint-plugin": "^2.23.0",
+    "@typescript-eslint/parser": "^2.23.0",
+    "electron": "^8.1.1",
+    "electron-builder": "^22.4.1",
+    "electron-webpack": "^2.7.4",
+    "electron-webpack-ts": "^3.2.0",
     "eslint": "^6.8.0",
     "eslint-config-prettier": "^6.10.0",
     "eslint-plugin-prettier": "^3.1.2",
-    "eslint-plugin-react": "^7.18.3",
-    "eslint-plugin-react-hooks": "^2.4.0",
-    "jsdom": "^16.2.0",
+    "eslint-plugin-react": "^7.19.0",
+    "eslint-plugin-react-hooks": "^2.5.0",
+    "file-loader": "^5.1.0",
+    "jsdom": "^16.2.1",
     "jsdom-global": "^3.0.2",
-    "mocha": "^7.0.1",
+    "mocha": "^7.1.0",
     "prettier": "^1.19.1",
-    "sinon": "^9.0.0",
-    "ts-node": "^8.6.2"
+    "sinon": "^9.0.1",
+    "ts-node": "^8.6.2",
+    "url-loader": "^3.0.0"
   }
 }
