@@ -1,4 +1,4 @@
-import { AppData } from "../persistence/app_data";
+import { AppData } from "../common/persistence/app_data";
 
 export interface ContentViewProps {
   appData: AppData;
