@@ -1,5 +1,5 @@
 import * as React from "react";
-import { Request } from "../request";
+import { Request } from "../common/request";
 import { ipcRequest } from "./communication";
 
 export const RequestContext = React.createContext<{
